refactor: PplainTime - phase 1 - util migration
